Added temporary_models.py, amended quiz page
